Move variations-related logic into utils
